Make poiDmg double

+++ battle/entity/Entity.java
@@ -1609,7 +1609,7 @@ public abstract class Entity extends AbEntity {
 			if (rst == 100) {
 				anim.getEff(INV);
 			} else {
-				int poiDmg = atk.getProc().POIATK.mult * (100 - rst) / 10000;
+				double poiDmg = atk.getProc().POIATK.mult * (100 - rst) / 10000.0;
 				damage += maxH * poiDmg;
 				basis.lea.add(new EAnimCont(pos, layer, effas().A_POISON.getEAnim(DefEff.DEF)));
 				basis.lea.sort(Comparator.comparingInt(e -> e.layer));
